Readded Timer.h, made able to compile on Linux machines, added one placeholder sound
I readded Timer.h since this seemed to be causing compiling issues, feel
free to modify this if removing Timer.h was intentional.  Additionally,
added Paddle.h and Paddle.cpp to the Makefile.am and configure.ac files
so that the assignment complies on the Linux machines.  Added SDL_mixer
sound clip of a snare drum when current block is broken; more sounds to
come (for some reason, the slime sound in resources didn't work).  I
haven't tested the push on the lab computers, will do that soon to make
sure this is a working push.